requirements.txt
pandas
tensorflow-gpu==2.0.0
tensorflow-addons
sklearn
numpy
loguru
pytest
python-dateutil
PyYAML
spacy
tensorflow-datasets
tqdm
pydot
graphviz
tensorboardX
jieba
prettytable
bentoml
sentencepiece